Add `src/mixins` folder (vue mixins refactor)
- create `src/mixins` folder
- deprecate notice for `src/app/gui/vue/vue.mixins.js`
- split and consolidate mixins definitions within a single folder
