+++ src/mixins/fields.js
@@ -1,0 +1,33 @@
+/**
+ * ORIGINAL SOURCE: src/app/gui/vue/vue.mixins.js@v3.6
+ */
+export default {
+  methods: {
+    getFieldService(){
+      if (this._fieldsService === undefined)
+        this._fieldsService = require('gui/fields/fieldsservice');
+      return this._fieldsService;
+    },
+    getFieldType(field) {
+      return this.getFieldService().getType(field);
+    },
+    isSimple(field){
+      return this.getFieldService().isSimple(field);
+    },
+    isLink(field){
+      return this.getFieldService().isLink(field);
+    },
+    isImage(field){
+      return this.getFieldService().isImage(field);
+    },
+    isPhoto(field){
+      return this.getFieldService().isPhoto(field);
+    },
+    isVue(field){
+      return this.getFieldService().isVue(field);
+    },
+    sanitizeFieldValue(value) {
+      return (Array.isArray(value) && !value.length) ? '' : value;
+    }
+  }
+};
