+++ src/mixins/index.js
@@ -1,0 +1,31 @@
+/**
+ * ORIGINAL SOURCE: src/app/gui/vue/vue.mixins.js@v3.6
+ */
+import autocompleteMixin from 'mixins/autocomplete';
+import fieldsMixin from 'mixins/fields';
+import mediaMixin from 'mixins/media';
+import geoMixin from 'mixins/geo';
+import resizeMixin from 'mixins/resize';
+import selectMixin from 'mixins/select';
+import select2Mixin from 'mixins/select2';
+import formInputsMixins from 'mixins/form-inputs';
+import widgetMixins from 'mixins/widget';
+import metadataMixin from 'mixins/metadata';
+import baseInputMixin from 'mixins/base-input';
+
+const mixins = {
+  autocompleteMixin,
+  fieldsMixin,
+  mediaMixin,
+  geoMixin,
+  resizeMixin,
+  selectMixin,
+  select2Mixin,
+  formInputsMixins,
+  widgetMixins,
+  metadataMixin,
+  baseInputMixin,
+}
+
+export { mixins };
+export default mixins;
