+++ src/mixins/media.js
@@ -1,0 +1,61 @@
+/**
+ * ORIGINAL SOURCE: src/app/gui/vue/vue.mixins.js@v3.6
+ */
+export default {
+  computed: {
+    filename() {
+      return this.value ? this.value.split('/').pop() : this.value;
+    }
+  },
+  methods: {
+    isMedia(value) {
+      if (value && typeof  value === 'object' && value.constructor === Object) return !!value.mime_type;
+      return false;
+    },
+    getMediaType(mime_type) {
+      let media = {
+        type: null,
+        options: {}
+      };
+      switch(mime_type) {
+        case 'image/gif':
+        case 'image/png':
+        case 'image/jpeg':
+        case 'image/bmp':
+          media.type = 'image';
+          break;
+        case 'application/pdf':
+          media.type = 'pdf';
+          break;
+        case 'video/mp4':
+        case 'video/ogg':
+        case 'video/x-ms-wmv':
+        case 'video/x-msvideo':
+        case 'video/quicktime':
+          media.type = 'video';
+          media.options.format = mime_type;
+          break;
+        case 'application/gzip':
+        case 'application/zip':
+          media.type = 'zip';
+          break;
+        case 'application/msword':
+        case 'application/vnd.oasis.opendocument.text':
+          media.type = 'text';
+          break;
+        case 'application/vnd.ms-office':
+        case 'application/vnd.oasis.opendocument.spreadsheet':
+          media.type = 'excel';
+          break;
+        case 'application/vnd.openxmlformats-officedocument.presentationml.presentation':
+        case 'application/vnd.ms-powerpoint':
+        case 'application/vnd.oasis.opendocument.presentation':
+          media.type = 'ppt';
+          break;
+        default:
+          media.type = 'unknow';
+      }
+      return media;
+    }
+  }
+};
